Replace lifetimes with reference counting
This library was originally written with a strategy of using Rust
lifetimes to ensure resources are dropped in the correct order. For
example, a `Context` must outlive a `Device`. This becomes a problem
when you want to store the objects together in a data structure.

    ```lang=rust
    struct DeviceManager {
        udev: libudev::Context,
        devices: Vec<libudev::Device<'self>,
    }
    ```

This won't work. There is no `'self` lifetime in Rust. There has been
discussion about how to enable this at the language level, but none of
the discussion has materialized into a solution, and it likely never
will. This just isn't how lifetimes are meant to be used.

Fortunately, libudev has built-in reference counting. This commit
replaces lifetime tracking with reference counting. By incrementing the
reference count in the `Clone` trait and decrementing the reference
count in the `Drop` trait, parent resource usage can be tracked by
simply holding onto copies.

This requires consideration of thread-safety. In particular, `Context`
objects can't be shared between threads. From `libudev(3)`:

> Furthermore, multiple different udev contexts can be used in parallel
> by multiple threads. However, a single context must not be accessed by
> multiple threads in parallel. The caller is responsible for providing
> suitable locking if they intend to use it from multiple threads.

Fortunately, Rust correctly infers `!Send` for `Context`, `Device`,
`Enumerator`, `Monitor`, and `MonitorSocket`, so clones of `Context`
nor other types that refer to a context internally may be sent to other
threads.