Bind Monitor with Context
Change construction of `Monitor` so that context is created along with it. This commit fixes issue #3
as monitor and context are now naturally kept together. In addition to that because MonitorSocket is
now the only owner of context (and is not cloneable) it can be safely send between threads.

Disadvantage of such approach is that application using not only monitor will implicitly create many
contexts.