Replace Context members with manual reference counting
Several types (`Device`, `Enumerator`, and `Monitor`) hold onto a
`Context`. While originally for tracking lifetime dependencies, it
turned into a technique for reference-counting parent resources. This
works, but is redundant with the libudev types that they wrap. libudev's
`udev_device`, `udev_enumerate`, and `udev_monitor` all hold references
to a `udev` context. It's redundant for this library's wrapper types to
hold a parallel reference.

Besides being redundant, it doubles the memory footprint of each type by
requiring two pointers instead of one. It also opens the possibility
that the `Context` and the `udev` members don't actually point to the
same object.

All of these issues are solved by replacing the `Context` member with
manual reference-counting. In each type's `Drop` trait, the `udev`
context can be obtained with `udev_*_get_udev()` and then its reference
count can be decremented.

I found one place where `Context` was being used for something other
than reference counting. That was to call
`context.device_from_syspath()` to create a `Device` from a `Devices`
iterator. This was a weird mix of high-level and low-level code, which
resulted in an unnecessary memory allocation by converting a C string to
`Path`, then to a `CString` (requiring a memory allocation), only to get
its pointer to pass back to a C function. By passing the C string
obtained from the enumerator directly to
`udev_device_new_from_syspath()`, all the transformations (and thus
superfluous memory allocation) can be skipped, and `Context` no longer
needs to be propagated to `Devices`.

`Devices` does, however, act on data borrowed from `Enumerator`, so it
requires a `PhantomData` marker to retain the correct borrow semantics.