Document !Send and !Sync for Context
While reviewing the reference-counting approach for thread-safety
issues, the issue of sharing a context between threads arose, because
the documentation in `libudev(3)` prohibits it:

> a single context must not be accessed by multiple threads in parallel

This means that `Context` and types that share a reference to a context
must be `!Send`. This commit adds this knowledge to the documentation so
that users of this library are aware of the limitation.